requirements.txt
# Core
Flask==3.0.0
SQLModel==0.0.14
python-dotenv==1.0.0
gunicorn==21.2.0

# QR Code Generation
qrcode[pil]==7.4.2

# i18n (optional for now)
# Flask-Babel==4.0.0

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0